test: base cypress.
