+++ generators/ionic/resources/base/cypress/fixtures/app.po.ts
@@ -1,0 +1,9 @@
+export class Page {
+  navigateTo(destination) {
+    return cy.visit(destination);
+  }
+
+  getTitle() {
+    return cy.title();
+  }
+}
