+++ generators/ionic/resources/base/cypress/integration/app.e2e-spec.ts
@@ -1,0 +1,19 @@
+import { Page } from '../fixtures/app.po';
+
+describe('App', () => {
+  let page: Page;
+
+  beforeEach(() => {
+    page = new Page();
+  });
+
+  describe('default screen', () => {
+    beforeEach(() => {
+      page.navigateTo('/');
+    });
+
+    it('should have the correct title', () => {
+      page.getTitle().should('include', 'Ionic App');
+    });
+  });
+});
