--- generators/ionic/resources/base/cypress/integration/spec.ts
@@ -1,7 +1,0 @@
-describe('My First Test', () => {
-  it('Visits the initial project page', () => {
-    cy.visit('/')
-    cy.contains('Welcome')
-    cy.contains('sandbox app is running!')
-  })
-})
