--- generators/ionic/resources/base/e2e/pages/app.po.ts
@@ -1,11 +1,0 @@
-import { browser } from 'protractor';
-
-export class Page {
-  navigateTo(destination) {
-    return browser.get(destination);
-  }
-
-  getTitle() {
-    return browser.getTitle();
-  }
-}
